% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{estRescueParams}
\alias{estRescueParams}
\title{Estimate Rescue Parameters}
\usage{
estRescueParams(
  sce,
  paramObj = NULL,
  sampleVariable = NULL,
  subjectVariable = NULL,
  groupVariable = NULL,
  timepointVariable = NULL,
  nonDEGs = NULL,
  cellParamsByCondition = FALSE
)
}
\arguments{
\item{sce}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
containing empirical data with a counts matrix in the \code{counts} slot and
cell level data (ex. sample/subject/timepoint labels) in the \code{colData}
slot.}

\item{paramObj}{\code{\link{RescueParams-class}} object with empty slots
for parameters which need to be estimated and parameters. If \code{NULL},
all parameter values will be estimated.}

\item{sampleVariable}{String denoting name of sample identifier variable in
the \code{colData} of \code{SingleCellExperiment} object. If \code{NULL},
data is assumed to contain only one sample parameters
\code{logLibFacVar}, \code{sampleFacVarMean}, \code{sampleFacVarSD},
\code{subjectFacVarMean}, and \code{subjectFacVarSD} parameters
cannot be estimated.}

\item{subjectVariable}{String denoting name of subject identifier variable in
the \code{colData} of \code{SingleCellExperiment} object. If \code{NULL},
parameters \code{nSubjsPerGroup}, \code{subjectFacVarMean}, and
\code{subjectFacVarSD} parameters cannot be estimated.}

\item{groupVariable}{String denoting name of group identifier variable in
the \code{colData} of \code{SingleCellExperiment} object. If \code{NULL}, a
single group design is assumed.}

\item{timepointVariable}{String denoting name of timepoint identifier
variable in the \code{colData} of \code{SingleCellExperiment} object.
If \code{NULL}, the number of timepoints for the simulated data is set to 2.}

\item{nonDEGs}{Vector containing names or row indices of genes to be used to
estimate sample and subject factor parameters. Using genes that are
differentially expressed across timepoints may lead to inaccurate estimates.}

\item{cellParamsByCondition}{Logical value indicating whether
\code{maxCellsPerSamp} and \code{minCellsPerSamp} should be estimated for
each condition (group/timepoint) or overall.}
}
\description{
Estimate parameters to be used to simulate repeated measures scRNA-Seq data
}
\details{
All parameters in \code{\link{RescueParams-class}} object can be
estimated/extracted from empirical data except \code{propDE} and
\code{deLogFC} which are both set to 0 (no differential expression) if not
set manually. If a paramObj is provided, only parameters with empty slots in
the object will be estimated.
}
\examples{
## Load the data
data("RecAM_sce")

## Subset data to 40 genes
RecAM_sce_small <- RecAM_sce[1:40, ]

## Estimate all parameters from the data
myParams_estimated <- estRescueParams(
  sce = RecAM_sce_small, paramObj = NULL,
  sampleVariable = "sampleID",
  subjectVariable = "subjectID",
  timepointVariable = "time",
  groupVariable = NULL, nonDEGs = NULL,
  cellParamsByCondition = F
)

## Example using Single timepoint data

## Subset the data to a single timepoint
RecAM_sce_single_time <- RecAM_sce_small[, RecAM_sce_small$time == "time0"]
## Create a params object with necessary parameters filled
myParams <- RescueParams(subjectFacVarMean = -4, subjectFacVarSD = 1, nSubjsPerGroup = 5)
## Estimate remaining parameters from the data
myParams_estimated <- estRescueParams(
  sce = RecAM_sce_single_time,
  paramObj = myParams,
  sampleVariable = "sampleID"
)

}
\author{
Elizabeth Wynn
}
