% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getRescueParam}
\alias{getRescueParam}
\title{Extract Parameters}
\usage{
getRescueParam(paramObj, paramName)
}
\arguments{
\item{paramObj}{Object of class \code{\link{RescueParams-class}}}

\item{paramName}{String containing parameter you would like to extract}
}
\value{
Specified parameter value
}
\description{
Extract individual parameters from RescueParams object
}
\examples{
## Create Parameter object
myParams <- RescueParams(nTimepoints = 2)

## Check Values
getRescueParam(myParams, "nTimepoints")

}
