% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getRescueSimParam}
\alias{getRescueSimParam}
\title{Extract Parameters}
\usage{
getRescueSimParam(paramObj, paramName)
}
\arguments{
\item{paramObj}{Object of class \code{\link{RescueSimParams-class}}}

\item{paramName}{String containing parameter you would like to extract}
}
\value{
Specified parameter value
}
\description{
Extract individual parameters from RescueSimParams object
}
\examples{
## Create Parameter object
myParams <- RescueSimParams(nTimepoints = 2)

## Check Values
getRescueSimParam(myParams, "nTimepoints")

}
