% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/data.R
=======
% Please edit documentation in R/RecAM_params.R
>>>>>>> bc9c617 (updated package)
\docType{data}
\name{RecAM_params}
\alias{RecAM_params}
\title{Example RNA-Seq data from paired observations}
\format{
\code{\link{RescueParams}} object with parameters estimated from
empirical dataset
}
\usage{
<<<<<<< HEAD
RecAM_params
=======
data(RecAM_params)
>>>>>>> bc9c617 (updated package)
}
\description{
Sample \code{\link{RescueParams}} object with
gene expression data for recruited airspace macrophage cells from
bronchoalveolar lavage samples samples collected from healthy adults.
The dataset from which parameters were estimated had 19,410 genes and 9,041
cells. Data was collected from five subjects at two timepoints per
subject.
}
\keyword{datasets}
