% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rescue-package.R
\docType{package}
\name{Rescue-package}
\alias{Rescue}
\alias{Rescue-package}
\title{Rescue: Repeated Measures Single Cell RNA-sequencing Data Simulation}
\description{
Simulate data from repeated measures scRNA-seq studies
}
\keyword{internal}
