% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RescueParams-class}
\alias{RescueParams-class}
\title{The RescueParams Class}
\description{
Class for holding parameters used to simulate data using \code{RESCUE}
}
\section{Slots}{

\describe{
\item{\code{nTimepoints}}{Number of timepoints (i.e. samples) per subject.
Holds a single numeric value >0 representing the number of timepoints for
all subjects, or a vector the same length as the number of total subjects
with each value representing the number of timepoints for a single subject.}

\item{\code{twoGroupDesign}}{Logical value indicating whether to simulate two groups
(ex. treatment and control group).}

\item{\code{nSubjsPerGroup}}{Number of subjects per group (if using two group design)
or number of total subjects (if only single group). Holds a single
numeric value >0 indicating the number of subjects per group, or a vector
 of two values with each value representing the subjects for one group (if
 using a two group design).}

\item{\code{maxCellsPerSamp}}{Maximum parameter used when drawing number of cells per
sample from a discrete uniform distribution. Holds a single numeric value
 >0 indicating the maximum number of cells per sample, or a vector with
 length equal to the number of conditions (group/timepoint combinations)
 with each value representing the maximum for a single condition, or a
 vector with length equal to the number of total samples with each value
 representing the maximum for a single sample.}

\item{\code{minCellsPerSamp}}{Minimum parameter used when drawing number of cells per
sample from a discrete uniform distribution. Holds a single numeric value >0
indicating the minimum number of cells per sample, or a vector with length
equal to the number of conditions (group/timepoint combinations) with each
value representing the minimum for a single condition, or a vector with
length equal to the number of total samples with each value representing
the minimum for a single sample.}

\item{\code{logLibMean}}{Mean library size (log scale) parameter. Used to draw
library size from a log-normal distribution. Holds a single value >0
indicating the mean library size on a log scale.}

\item{\code{logLibSD}}{Library size standard deviation (log scale) parameter.
Used to draw library size from a log-normal distribution. Holds a
single value >=0 indicating the standard deviation of library size on
a log scale.}

\item{\code{logLibFacVar}}{Variance used for drawing sample-level multiplicative
factors which give different library size distributions for each sample.
Larger values result in larger variation in library size distributions by
sample. Holds a single value >=0 with 0 indicating no difference in the
library size distributions by sample.}

\item{\code{exprsMean}}{Gene-specific mean expression value representing the average
expression of each gene in the dataset. Holds a vector of numeric values >=0
with length equal to desired number of genes in the simulated data where each
value indicates the average expression for a single gene.}

\item{\code{dispersion}}{Gene-specific dispersion value representing the variation in
expression for each gene in the dataset. Holds a vector of numeric values >0
with length equal to desired number of genes in the simulated data where each
value indicates the dispersion for a single gene.}

\item{\code{sampleFacVarMean}}{Mean used for drawing variance (log-scale) of
sample-level multiplicative factors. Larger values result in
more between-sample variation.}

\item{\code{sampleFacVarSD}}{Standard deviation used for drawing variance (log-scale)
of sample-level multiplicative factors. Larger values result in more
variation in amount of between-sample variation across different genes.
Must be a value >=0}

\item{\code{subjectVarMean}}{Mean used for drawing variance (log-scale) of
subject-level multiplicative factors. Larger values result in larger
between-subject variation.}

\item{\code{subjectFacVarSD}}{Standard deviation used for drawing variance
(log-scale) of subject-level multiplicative factors. Larger values result in
more variation in amount of between-subject variation across different genes.
Must be a value >=0.}

\item{\code{propDE}}{Proportion of genes differentially expressed between
timepoints/groups. Must be a numeric value between 0 and 1.}

\item{\code{deLogFC}}{Fold change values used for differentially expressed genes.
A single numeric value >=0 indicating the positive log-fold change for
differentially expressed genes (log2 fold change values will be simulated
evenly to be positive or negative) or a vector of values (positive of
negative) to draw log2 fold change values from.}
}}

\author{
Elizabeth Wynn
}
