% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RescueParams}
\alias{RescueParams}
\title{Generate RescueParams object}
\usage{
RescueParams(...)
}
\arguments{
\item{...}{Any parameter name followed by initial valus}
}
\value{
An object of class \code{RescueParams}
}
\description{
Generate RescueParams object
}
\examples{
## Create params object
myParams <- RescueParams()

## Create parameter object with nTimepoints and nSubjsPerGroup pre-set
myParams <- RescueParams(nTimepoints = 2, nSubjsPerGroup = 5)

}
\author{
Elizabeth Wynn
}
