% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rescue-package.R
\docType{package}
\name{rescueSim-package}
\alias{rescueSim}
\alias{rescueSim-package}
\title{rescueSim: Repeated Measures Single Cell RNA-sequencing Data Simulation}
\description{
A package for simulating data from repeated measures scRNA-seq studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ewynn610/rescueSim}
  \item \url{https://ewynn610.github.io/rescueSim/}
  \item Report bugs at \url{https://github.com/ewynn610/rescueSim/issues}
}

}
\author{
\strong{Maintainer}: Elizabeth Wynn \email{wynne@njhealth.org}

}
\keyword{internal}
