% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simRescueData}
\alias{simRescueData}
\title{Simulate Repeated Measures scRNA-seq Data}
\usage{
simRescueData(paramObj)
}
\arguments{
\item{paramObj}{\code{\link{RescueParams-class}} object with no empty slots.}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object with
the data in the following slots

\describe{
\item{\code{counts}}{Matrix of raw counts with genes represented by rows and
cells represented by columns.}
    \item{\code{colData}}{
        \describe{
            \item{sampleID}{Sample identifier}
            \item{subjectID}{Subject identifier}
            \item{time}{Timepoint identifier}
            \item{group}{Group identifier}
        }
    }
    \item{\code{rowData}}{
        \describe{
            \item{deLogFC}{Log2 fold change between first and last timepoints
            or groups, or, if multi-timepoint and multi-group data, Log2 fold
             change between the last timepoint and group vs. other
             timepoint/group combinations}
        }
    }

}
}
\description{
Simulate Repeated Measures scRNA-seq Data using the RESCUE method
}
\examples{
# Read in parameter object
data("RecAM_params")

# Simulate data
simDat=simRescueData(RecAM_params)

# Examine
library(SingleCellExperiment)
counts(simDat)[1:5, 1:5]

}
