% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{updateRescueParams}
\alias{updateRescueParams}
\title{Update RescueParams object}
\usage{
updateRescueParams(paramObj, paramValues)
}
\arguments{
\item{paramObj}{Object of class \code{\link{RescueParams-class}}}

\item{paramValues}{List of parameter values with list names are the parameter
names}
}
\value{
Object of class \code{\link{RescueParams-class}} with specified parameters updated
}
\description{
Manually update parameters in RescueParams object
}
\examples{
## Create Parameter object
myParams <- RescueParams()

## Update nTimepoints and nSubjsPerGroup
myParams <- updateRescueParams(
  paramObj = myParams,
  paramValues = list(
    nTimepoints = 3,
    nSubjsPerGroup = 10
  )
)
## Check Values
getRescueParam(myParams, "nTimepoints")
getRescueParam(myParams, "nSubjsPerGroup")

}
