% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{updateRescueSimParams}
\alias{updateRescueSimParams}
\title{Update RescueSimParams object}
\usage{
updateRescueSimParams(paramObj, paramValues)
}
\arguments{
\item{paramObj}{Object of class \code{\link{RescueSimParams-class}}}

\item{paramValues}{List of parameter values with list names are the parameter
names}
}
\value{
Object of class \code{\link{RescueSimParams-class}} with specified parameters updated
}
\description{
Manually update parameters in RescueSimParams object
}
\examples{
## Create Parameter object
myParams <- RescueSimParams()

## Update nTimepoints and nSubjsPerGroup
myParams <- updateRescueSimParams(
  paramObj = myParams,
  paramValues = list(
    nTimepoints = 3,
    nSubjsPerGroup = 10
  )
)
## Check Values
getRescueSimParam(myParams, "nTimepoints")
getRescueSimParam(myParams, "nSubjsPerGroup")

}
