% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rescue-package.R
\docType{package}
\name{RESCUE-package}
\alias{RESCUE}
\alias{RESCUE-package}
\title{RESCUE: Repeated Measures Single Cell RNA-sequencing Data Simulation}
\description{
A package for simulating data from repeated measures scRNA-seq studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ewynn610/Rescue}
  \item Report bugs at \url{https://github.com/ewynn610/Rescue/issues}
}

}
\author{
\strong{Maintainer}: Elizabeth Wynn \email{wynne@njhealth.org}

}
\keyword{internal}
