% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RescueSimParams}
\alias{RescueSimParams}
\title{Generate RescueSimParams object}
\usage{
RescueSimParams(deLogFC = 0, propDE = 0, ...)
}
\arguments{
\item{...}{Any parameter name followed by initial valus}
}
\value{
An object of class \code{RescueSimParams}
}
\description{
Generate RescueSimParams object
}
\examples{
## Create params object
myParams <- RescueSimParams()

## Create parameter object with nTimepoints and nSubjsPerGroup pre-set
myParams <- RescueSimParams(nTimepoints = 2, nSubjsPerGroup = 5)

}
\author{
Elizabeth Wynn
}
