% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simRescueData}
\alias{simRescueData}
\title{Simulate Repeated Measures scRNA-seq Data}
\usage{
simRescueData(paramObj)
}
\arguments{
\item{paramObj}{\code{\link{RescueParams-class}} object with no empty slots.}
}
\value{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object with
the data in the following slots

\describe{
\item{\code{counts}}{Matrix of raw counts with genes represented by rows and
cells represented by columns.}
    \item{\code{colData}}{
        \describe{
            \item{sampleID}{Sample identifier}
            \item{subjectID}{Subject identifier}
            \item{time}{Timepoint identifier}
            \item{group}{Group identifier}
        }
    }
\item{\code{rowData}}{
  \describe{
    \item{deLogFC}{A \code{DataFrame} containing log2 fold change information for each gene.
    Each column corresponds to a non-reference experimental condition
    (e.g., \code{"time1"}, \code{"group1"}, \code{"time1_group1"}, etc.), and values represent
    gene-level log2 fold changes relative to the baseline condition
    (\code{"time0"}, \code{"group0"}, or \code{"time0_group0"} depending on the design).
    The reference condition itself is not included as a column.}
  }
}

}
}
\description{
Simulate Repeated Measures scRNA-seq Data using the RESCUE method
}
\examples{
# Read in parameter object
data("RecAM_params")

# Simulate data
simDat=simRescueData(RecAM_params)

# Examine
library(SingleCellExperiment)
counts(simDat)[1:5, 1:5]

}
